Urban milestones 8-15, the POI freeze, and Vulkan answered by measurement
Seven urban morphology modules land at once -- radial (m8), water (m9),
fortify (m10), cleanup (m11), districts (m13), amenities (m14) and
hinterland (m15). cartalith-urban goes from ten declared modules to
seventeen and from 119 to 258 passing tests. Each milestone carries a
golden that reproduces the reference exactly; all three spot-checked
constants were mutation-verified to prove the goldens can actually fail.
fortify.rs ports FDLIBM's __ieee754_acos with bit-pattern constants,
because the scope document measured Math.acos disagreeing with f64::acos
on 544 of 60 000 arguments and corner_cut compares the result against a
threshold. Milestone 16 is now unblocked -- STATUS.md filed it "blocked
by definition on 8-15" and all eight exist.

The owner reported the point-of-interest pass freezing and not drawing.
Three causes, not two. landmark_run() was synchronous on the main thread:
0 frames served during a 1 225 ms pass, against 255 during a generate()
doing four times the work. Nothing pushed placements at the map --
refresh_annotations() is MapOverlay._landmarks' only writer and no
landmark path called it. And moving the pass to a Thread produced
"attempted to access binding from different thread" out of godot-ffi,
because without experimental-threads every Dictionary/Array/GString op
routes through ensure_main_thread(). A worker-thread #[func] must be
primitives-in, primitive-out; landmark_run now returns bool and
landmark_last_run builds the reply on the main thread. Threading it also
removed the accidental guard the freeze provided, so civ_recompute,
grid_size and get_factions gained the generating check they lacked.
4.14 s -> 0.39 s at the shipping default, and off the main thread.

Vulkan, DirectX and RenderingDevice all answered no, the first by
measurement rather than caution: on the owner's RX 7800 XT,
forward_plus/vulkan loses the GPU device during generate 3 runs of 3,
forward_plus/d3d12 segfaults, gl_compatibility is clean. DX12 needed no
work -- COMPUTE_BACKENDS already unions it and masks out only OpenGL,
itself a bisected signal-11 fix. WGPU_BACKEND was inert, so the
Vulkan-first preference could never be tested; it now reads the
environment with the OpenGL mask applied after, and DX12 was exercised
for the first time (59/59, an identical test list to Vulkan).

Also: a software rasterizer could be selected by default, because
force_fallback_adapter: false declines to restrict rather than excluding
CPU adapters -- the doc comment asserted the opposite. No log backend was
registered anywhere, which made the Android "zero wgpu lines in logcat"
pass condition unfalsifiable. Vault milestones 2 and 3, with a cross-world
snapshot leak the verify pass caught and a probe that now pins it. R4/R5/
R7/R8, Rayon over road_dijkstra's independent sources (ordering proved by
mutation), _civPlaceSmelting and _civSaltAccess with a new golden suite.

Documented-but-false claims corrected in place: the commit row both
STATUS.md and OUTSTANDING_WORK.md led with had already shipped in fd9de7c;
the GPU determinism flake was closed by 803b725; a gl_compatibility
rationale does exist, in the godot-shell skill committed alongside
project.godot.

+++ cartalith-native/crates/cartalith-engine/src/import.rs
@@ -200,7 +200,7 @@ pub fn infer_tectonics(field: Vec<f32>, p: &WorldParams) -> WorldState {
 
     // `baseField` = per-cell plate base, then the same 0.35x blur the
     // forward substrate applies (reference line 6767).
-    let crust_field: Vec<f32> = plate_id.iter().map(|&pi| plates[pi].base as f32).collect();
+    let crust_field: Vec<f32> = plate_id.iter().map(|&pi| plates[pi as usize].base as f32).collect();
     let _base_field = gauss_blur(&crust_field, (p.tect.blur_r * 0.35).max(2.0), gw, gh, world);
 
     // ---- forward stages, reused verbatim (reference HTML lines 6769-6776) ----
